commit 10: Abschnitte in LazyColumn bei CourseDetails fertig

+++ app/src/main/java/com/example/probeui/classes/Lection.kt
@@ -5,7 +5,7 @@ data class Lection(
     val mCourseId: Int,
     val mLectionName: String,
     val mType: ELectionType,
-    val mSection: Short,
+    val mSection: Int,
     val mPositionInSection: Short,
     val mDuration: String,
     val mUrl: String?,
